requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dj-database-url==1.2.0
gunicorn==20.1.0
